Obtainable Safari Balls...?
Here's one for all the people that wanted another wacky interaction with the shady salesman!

He stands outside the Safari Zone selling his own Safari Balls. They act the same way, they're called the same thing, but they're looking a little off...hey, is that paint peeling off?

Anyway, this acts as a way to get Safari Balls outside of the Safari Zone, like in SwSh! Whether or not you wanna pretend that they're real, it's still neat that he's selling them to you for a good price.