Restore unreferenced Porygon PC
This was deleted when I was sprucing up the final Silph Co. floor, but I decided to restore it instead.

Because of the weird way PCs work, I made it so the text is a literal bg event on a monitor, true to the text. I think it works fine.

The text was changed to use `prompt` instead of `done` to fix text scrolling issues.