Safari Zone tweaks
Added some Cubones since it's featured in the Safari Zone in Yellow and Pokemon Origins. Changed the second Chansey's level to 7 for consistency.

Also fixed an error on the Safari Zone Area 1 sign that said Centre Area, which was fixed in FRLG. Thanks to Lyra made a website for pointing this out!