Facility floor tile change
This was a thing I made back in March, then got sidetracked by the Yoshi April Fool's thing, accidentally deleted and forgot about but now I've reimplemented. Oops.

This changes the tileset of Rocket Hideout, Silph Co. and Cinnabar Mansion to be much easier on the eyes, especially when holding the run button.